refactor: count 계열 JPA 메서드의 반환 타입 변경
- PK 타입과 동일한 `Long`으로 변경

+++ src/main/kotlin/com/barostartbe/domain/assignment/repository/AssignmentRepository.kt
@@ -55,5 +55,5 @@ interface AssignmentRepository : JpaRepository<Assignment, Long> {
         ) streaks
     """, nativeQuery = true
     )
-    fun findMaxConsecutivePerfectDays(@Param("menteeId") menteeId: Long): Int
+    fun findMaxConsecutivePerfectDays(@Param("menteeId") menteeId: Long): Long
 }
